:pencil: Create ducumentation for RecursiveWhile.kt
